+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -2,8 +2,8 @@
 
 ## Version 4.1 (dev)
 
-* Added `StatementFilter` and `ReferencedStatementFilter`
 * Added `StatementList::filter`
+* Added `StatementFilter` and `ReferencedStatementFilter`
 * Added `LabelsProvider`, `DescriptionsProvider` and `AliasesProvider`
 
 ## Version 4.0 (2015-07-28)