+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -1,10 +1,11 @@
 # Wikibase DataModel release notes
 
-## Version 4.1 (dev)
+## Version 4.1 (2015-08-04)
 
 * Added `StatementList::filter`
 * Added `StatementFilter` and `ReferencedStatementFilter`
 * Added `LabelsProvider`, `DescriptionsProvider` and `AliasesProvider`
+* Added `FingerprintHolder`
 
 ## Version 4.0 (2015-07-28)